misc: compiler warning fixes, refs #33

+++ src/qcommon/vm_interpreted.c
@@ -557,8 +557,8 @@ nextInstruction2:
 #endif
 				*(int *)&image[programStack + 4] = -1 - programCounter;
 
-//VM_LogSyscalls( (int *)&image[ programStack + 4 ] );
-				r = vm->systemCall((int *)&image[programStack + 4]);
+				//VM_LogSyscalls( (int *)&image[ programStack + 4 ] );
+				r = vm->systemCall((intptr_t *)&image[programStack + 4]);
 
 #ifdef DEBUG_VM
 				// this is just our stack frame pointer, only needed
